movies

+++ movies/views.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render, redirect
-from .models import Movie
-from .forms import MovieForm
+from .models import Movie, Comment
+from .forms import MovieForm, CommentForm
 
 
 def index(request):
@@ -15,7 +15,10 @@ def create(request):
     if request.method == 'POST':
         form = MovieForm(request.POST)
         if form.is_valid():
-            movie = form.save()
+            movie = form.save(commit=False)
+            movie.user = request.user
+            movie.save()
+
             return redirect('movies:detail', movie.pk)
     else:
         form = MovieForm()
@@ -27,20 +30,28 @@ def create(request):
 
 def detail(request, pk):
     movie = Movie.objects.get(pk=pk)
+    comment_form = CommentForm()
+    comments = movie.comment_set.all()
     context = {
         'movie': movie,
+        'comment_form' : comment_form,
+        'comments' : comments,
     }
     return render(request, 'movies/detail.html', context)
 
 def update(request, pk):
     movie = Movie.objects.get(pk=pk)
-    if request.method == 'POST':
-        form = MovieForm(request.POST, instance=movie)
-        if form.is_valid():
-            form.save()
+    if request.user == movie.user:
+        if request.method == 'POST':
+            form = MovieForm(request.POST, instance=movie)
+            if form.is_valid():
+                form.save()
             return redirect('movies:detail', movie.pk)
+        else:
+           form = MovieForm(instance=movie)
+
     else:
-        form = MovieForm(instance=movie)
+        return redirect('movies:detail', movie.pk)
     context = {
         'movie': movie,
         'form': form,
@@ -49,14 +60,38 @@ def update(request, pk):
 
 def delete(request, pk):
     movie = Movie.objects.get(pk=pk)
-    movie.delete()
-    return redirect('movies:index')
+    if request.user == movie.user:
+        movie.delete()
+        return redirect('movies:index')
+    return redirect('movies:detail', movie.pk)
 
-def comment_create(request):
-    pass
+def comments_create(request, pk):
+    if request.user.is_authenticated:
+        movie = Movie.objects.get(pk=pk)
+        comment_form = CommentForm(request.POST)
+        if comment_form.is_valid():
+            comment = comment_form.save(commit=False)
+            comment.movie = movie
+            comment.user = request.user
+            comment.save()
+        return redirect('movies:detail', movie.pk)
+    
+    return redirect('accounts:login')
 
-def comment_delete(request, movie_pk, comment_pk):
-    pass
+def comments_delete(request, movie_pk, comment_pk):
+    if request.user.is_authenticated:
+        comment = Comment.objects.get(pk=comment_pk)
+        if request.user == comment.user:
+            comment.delete()
+    return redirect('movies:detail' , movie_pk)
 
-def likes(request):
-    pass
+
+def likes(request, pk):
+    if request.user.is_authenticated:
+        movie = Movie.objects.get(pk=pk)
+        if movie.like_users.filter(pk=request.user.pk).exists():
+            movie.like_users.remove(request.user)
+        else:
+            movie.like_users.add(request.user)
+        return redirect('movies:detail', pk)
+    return redirect('accounts:login')
